+++ sentry-apollo-5/src/test/java/io/sentry/util/Apollo5PlatformTestManipulator.kt
@@ -1,0 +1,7 @@
+package io.sentry.util
+
+object Apollo5PlatformTestManipulator {
+  fun pretendIsAndroid(isAndroid: Boolean) {
+    Platform.isAndroid = isAndroid
+  }
+}